dir structure redone
