+++ scripts/check_frames
@@ -1,1 +1,3 @@
+#!/usr/bin/env bash
+
 ffmpeg -i $1 -map 0:v:0 -c copy -f null -y /dev/null 2>&1 | grep -Eo 'frame= *[0-9]+ *' | grep -Eo '[0-9]+' | tail -1
