+++ scripts/check_timestamps
@@ -1,3 +1,5 @@
+#!/usr/bin/env conda run -n setup python
+
 "Visualize delays between frames and frame drops."
 import glob
 import pickle
